requirements.txt
Flask==3.0.0
Pillow>=10.0.0
werkzeug==3.0.1
